backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
bcrypt==4.0.1
httpx==0.25.0
email-validator==2.1.0
anthropic==0.18.1
